+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `TypeId` uses `strum::FromRepr`
 - Removed C path utils
 - Changed some naming convention
+- Removed all `NulError` from error types
 
 ### Removed
 - Cleaned up unused code